+++ source/release-notes.md
@@ -892,4 +892,6 @@
 
 17.06.2019
 * maintenance-plan: option to allow/block maintenance operations on a non-readable AlwaysOn secondary replica
+* maintenance-plan: add time limit option for the consistency tasks in maintenance plan
 * health-check: collect secondary replica readable state in [health-check].[statsDatabaseAlwaysOnDetails] table
+* add time limit for executing internal jobs (dbo.usp_jobQueueExecute)